Invite 작업 및 페이징 테스트 완료
- Invite 내용이 없을 때 어디서 막을지 결정 후 작업

- DB 정리

- Post 분리 시작

+++ InstaClone/src/main/java/com/skhkim/instaclone/chatting/repository/ChatRoomRepository.java
@@ -18,5 +18,5 @@ public interface ChatRoomRepository extends JpaRepository<ChatRoom, String> {
     @Modifying
     @Transactional
     @Query("UPDATE ChatRoom cr SET cr.userNum = cr.userNum + :addNum WHERE cr.roomId = :roomId")
-    void updateUserNumByRoomId(Long roomId, Long addNum);
+    void updateUserNumByRoomId(Long roomId, Integer addNum);
 }
